requirements.txt
Django
django-location-field